requirements.txt
pyperclip
google-generativeai
pynput
pystray
customtkinter